refactor(beacon-api-types): move types around clean up modules
