+++ cosmwasm/ibc-union/lightclient/ethereum/src/errors.rs
@@ -1,5 +1,5 @@
 use beacon_api_types::Slot;
-use cosmwasm_std::{StdError, VerificationError};
+use cosmwasm_std::StdError;
 use ibc_union_light_client::IbcClientError;
 use unionlabs::primitives::{H256, U256};
