+++ lib/beacon-api-types/src/fork.rs
@@ -1,4 +1,4 @@
-use crate::Version;
+use crate::custom_types::Version;
 
 #[derive(Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "ssz", derive(ssz::Ssz))]
