+++ lib/beacon-api-types/src/light_client_update.rs
@@ -2,7 +2,8 @@ use unionlabs::primitives::H256;
 
 use crate::{
     consts::{floorlog2, FINALIZED_ROOT_INDEX, NEXT_SYNC_COMMITTEE_INDEX},
-    LightClientHeader, Slot, SyncAggregate, SyncCommittee,
+    slot::Slot,
+    LightClientHeader, SyncAggregate, SyncCommittee,
 };
 
 pub type NextSyncCommitteeBranch = [H256; floorlog2(NEXT_SYNC_COMMITTEE_INDEX)];
