+++ lib/ethereum-light-client-types/src/consensus_state.rs
@@ -1,4 +1,4 @@
-use beacon_api_types::Slot;
+use beacon_api_types::slot::Slot;
 use unionlabs::primitives::{H256, H384};
 
 #[derive(Debug, Clone, PartialEq)]
