+++ lib/ethereum-light-client-types/src/header.rs
@@ -19,8 +19,8 @@ pub struct Header {
 #[cfg(test)]
 mod tests {
     use beacon_api_types::{
-        execution_payload_header::ExecutionPayloadHeader, BeaconBlockHeader, LightClientHeader,
-        Slot, SyncAggregate, SyncCommittee,
+        execution_payload_header::ExecutionPayloadHeader, slot::Slot, BeaconBlockHeader,
+        LightClientHeader, SyncAggregate, SyncCommittee,
     };
     use unionlabs::{
         encoding::{Bincode, Json},
