+++ voyager/modules/client-bootstrap/ethereum/src/main.rs
@@ -5,7 +5,7 @@ use alloy::{
     rpc::types::BlockTransactionsKind,
 };
 use beacon_api::client::BeaconApiClient;
-use beacon_api_types::{PresetBaseKind, Slot};
+use beacon_api_types::{chain_spec::PresetBaseKind, slot::Slot};
 use ethereum_light_client_types::{ClientState, ConsensusState};
 use jsonrpsee::{
     core::{async_trait, RpcResult},
